refactor: Renamed and refactored the Register User DTO to the auth folder
